Extract Neo4j concept model in a separate package under ontology
